refactor(memory_bridge): promote to directory, split tests (#1266)
- src/memory_bridge.rs: 882 → src/memory_bridge/mod.rs (389 LOC)
- New src/memory_bridge/tests.rs (493 LOC, follow-up split needed)

Pure refactor: 34 tests pass.
